requirements.txt
beautifulsoup4==4.12.3
dash==2.17.1
dash_bootstrap_components==1.6.0
Flask==3.1.0
numpy==2.1.3
pandas==2.2.3
plotly==5.22.0
python_dateutil==2.9.0.post0
Requests==2.32.3
scikit_learn==1.4.2
tensorflow==2.18.0
tensorflow==2.17.0
tensorflow_intel==2.18.0
tensorflow_intel==2.17.0
tqdm==4.66.4
